+++ tardis/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this NixOS configuration will be documented in this file.
 
+## 2025-12-24
+
+### Added
+- Enabled envfs to provide FHS-compatible paths for legacy shebangs (e.g., /bin/bash)
+
+### Changed
+- nixos/configuration.nix: Enabled services.envfs
+- home-manager/zsh.nix: Updated osup to use the flake target
+- nixos/configuration.nix: Disabled xpadneo
+- nixos/configuration.nix: Enabled xone for Xbox controllers
+- nixos/configuration.nix: Temporarily disabled xone due to xpad-noone build failure on Linux 6.18; retry later
+
 ## 2025-11-04
 
 ### Added
@@ -19,5 +31,4 @@ All notable changes to this NixOS configuration will be documented in this file.
   - Commented out sunshine.nix import in home-manager/home.nix
   - Commented out sunshine package in home-manager/home.nix
   - Commented out Sunshine firewall rules in nixos/network.nix (TCP/UDP ports 47984-48010)
-  - Commented out all configuration in home-manager/sunshine.nix with "DISABLED" marker
-
+- Commented out all configuration in home-manager/sunshine.nix with "DISABLED" marker